backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
# google-generativeai==0.3.1
pytest==7.4.3
httpx==0.25.1
pydantic
google-genai